cleanup settings
